+++ FLEA/FLEA/Db/ActiveRecord.php
@@ -25,21 +25,21 @@ class ActiveRecord
      *
      * @var array
      */
-    public $_aggregation = [];
+    public $aggregation = [];
 
     /**
      * 用于完成数据库操作的 TableDataGateway 继承类
      *
      * @var \FLEA\Db\TableDataGateway
      */
-    public $_table;
+    public $table;
 
     /**
      * 该对象的主键属性名
      *
      * @var string
      */
-    public $_idname;
+    public $idname;
 
     /**
      * 指示该对象是否已经初始化
@@ -53,7 +53,7 @@ class ActiveRecord
      *
      * @var array
      */
-    public $_mapping = false;
+    public $mapping = false;
 
     /**
      * 继承类必须覆盖此静态函数
@@ -62,7 +62,7 @@ class ActiveRecord
      *
      * @return array
      */
-    static function define(): array
+    public static function define(): array
     {
     }
 
@@ -95,29 +95,29 @@ class ActiveRecord
 
         $objid = "{$myclass}_tdg";
         if (\FLEA::isRegistered($objid)) {
-            $this->_table = \FLEA::registry($objid);
+            $this->table = \FLEA::registry($objid);
         } else {
             // 使用 Composer PSR-4 自动加载
             if (!class_exists($tableClass, true)) {
                 throw new \FLEA\Exception\ExpectedClass($tableClass);
             }
-            $this->_table = new $tableClass(['skipCreateLinks' => true]);
-            \FLEA::register($this->_table, $objid);
+            $this->table = new $tableClass(['skipCreateLinks' => true]);
+            \FLEA::register($this->table, $objid);
         }
 
         if (!empty($options['propertiesMapping'])) {
-            $this->_mapping = [
+            $this->mapping = [
                 'p2f' => $options['propertiesMapping'],
                 'f2p' => array_flip($options['propertiesMapping']),
             ];
-            $this->_idname = $this->_mapping['f2p'][$this->_table->primaryKey];
+            $this->idname = $this->mapping['f2p'][$this->table->primaryKey];
         } else {
-            $this->_mapping = ['p2f' => [], 'f2p' => []];
-            foreach ($this->_table->meta as $field) {
-                $this->_mapping['p2f'][$field['name']] = $field['name'];
-                $this->_mapping['f2p'][$field['name']] = $field['name'];
+            $this->mapping = ['p2f' => [], 'f2p' => []];
+            foreach ($this->table->meta as $field) {
+                $this->mapping['p2f'][$field['name']] = $field['name'];
+                $this->mapping['f2p'][$field['name']] = $field['name'];
             }
-            $this->_idname = $this->_table->primaryKey;
+            $this->idname = $this->table->primaryKey;
         }
 
         if (!isset($options['aggregation']) || !is_array($options['aggregation'])) {
@@ -153,9 +153,9 @@ class ActiveRecord
                 $link['assocForeignKey'] = $define['assocForeignKey'] ?? null;
             }
 
-            $this->_table->createLink($link, $define['mappingType']);
-            $define['link'] =& $this->_table->getLink($link['mappingName']);
-            $this->_aggregation[$offset] = $define;
+            $this->table->createLink($link, $define['mappingType']);
+            $define['link'] =& $this->table->getLink($link['mappingName']);
+            $this->aggregation[$offset] = $define;
         }
     }
 
@@ -166,7 +166,7 @@ class ActiveRecord
      */
     public function load($conditions): void
     {
-        $row = $this->_table->find($conditions);
+        $row = $this->table->find($conditions);
         if (is_array($row)) { $this->attach($row); }
     }
 
@@ -176,7 +176,7 @@ class ActiveRecord
     public function save(): void
     {
         $row =& $this->toArray();
-        $this->_table->save($row);
+        $this->table->save($row);
     }
 
     /**
@@ -184,7 +184,7 @@ class ActiveRecord
      */
     public function delete(): void
     {
-        $this->_table->removeByPkv($this->getId());
+        $this->table->removeByPkv($this->getId());
     }
 
     /**
@@ -194,7 +194,7 @@ class ActiveRecord
      */
     public function setId($id): void
     {
-        $this->{$this->_idname} = $id;
+        $this->{$this->idname} = $id;
     }
 
     /**
@@ -204,7 +204,7 @@ class ActiveRecord
      */
     public function getId()
     {
-        return $this->{$this->_idname};
+        return $this->{$this->idname};
     }
 
     /**
@@ -215,7 +215,7 @@ class ActiveRecord
     public function toArray(): array
     {
         $arr = [];
-        foreach ($this->_mapping['p2f'] as $prop => $field) {
+        foreach ($this->mapping['p2f'] as $prop => $field) {
             $arr[$field] = $this->{$prop};
         }
         return $arr;
@@ -228,13 +228,13 @@ class ActiveRecord
      */
     public function attach(array &$row): void
     {
-        foreach ($this->_mapping['f2p'] as $field => $prop) {
+        foreach ($this->mapping['f2p'] as $field => $prop) {
             if (isset($row[$field])) {
                 $this->{$prop} = $row[$field];
             }
         }
 
-        foreach ($this->_aggregation as $define) {
+        foreach ($this->aggregation as $define) {
             $mn = $define['link']->mappingName;
             if (!isset($row[$mn])) { continue; }
             if ($define['link']->oneToOne) {