+++ FLEA/FLEA/Db/TableLink/HasOneLink.php
@@ -25,7 +25,7 @@ class HasOneLink extends TableLink
      *
      * @var boolean
      */
-    public $oneToOne = true;
+    public bool $oneToOne = true;
 
     /**
      * 返回用于查询关联表数据的SQL语句
@@ -34,14 +34,14 @@ class HasOneLink extends TableLink
      *
      * @return string
      */
-    public function getFindSQL($in)
+    public function getFindSQL(string $in): string
     {
         if (!$this->init) { $this->init(); }
         $fields = $this->qforeignKey . ' AS ' . $this->mainTDG->pka . ', ' . $this->assocTDG->qfields($this->fields);
 
         $sql = "SELECT {$fields} FROM {$this->assocTDG->qtableName} ";
 
-        return parent::_getFindSQLBase($sql, $in);
+        return parent::getFindSQLBase($sql, $in);
     }
 
     /**
@@ -52,12 +52,12 @@ class HasOneLink extends TableLink
      *
      * @return boolean
      */
-    function saveAssocData(array &$row, $pkv): bool
+    public function saveAssocData(array &$row, $pkv): bool
     {
         if (!$this->init) { $this->init(); }
 
         $row[$this->foreignKey] = $pkv;
-        return $this->_saveAssocDataBase($row);
+        return $this->saveAssocDataBase($row);
     }
 
     /**